Refactor a little
Trying to pull this apart a little. It seems like it'd
be better replaced by a leex & yecc parser. We'll leave
it as is an simply rewrite the code later. The tests are
also very CouchDB specific. I don't know why this was
pulled into a separate repository when it's pretty much
impossible to use generically withouth modification. That's
a rant for another time (shame on you CouchDB).

The big puzzle will be moving the application environment
mapping stuff somewhere. I'll leave that up to a different
app (called dummy, node and cluster management stuff).